add lunar date

+++ src/com/asys/weather/util/Config.java
@@ -14,6 +14,6 @@ public class Config {
 	public static final float HIDE_SHOW_PAGE_ANIMATION_TIME = 0.2f;
 	public static final float HIDE_SHOW_PAGE_ANIMATION_DETA = 0.1f;
 
-	public static final long HTTP_TIMEOUT = 10000;
+	public static final int HTTP_TIMEOUT = 10000;
 
 }
